// packages/frontend/src/types/game.ts
export interface Vec3 {
  x: number;
  y: number;
  z: number;
}

export enum ResourceType {
  Wood = 'wood',
  Stone = 'stone',
  Berry = 'berry',
}

export interface Resource {
  id: string;
  type: ResourceType;
  position: Vec3;
  harvested: boolean;
}